Move config-related spawn types to new module
